Closes #6,closes #7,closes #8, closes #9; Implements screening and room feature.

+++ ticket-service/src/main/java/com/epam/training/ticketservice/core/movie/persistence/repository/MovieRepository.java
@@ -7,8 +7,9 @@ import org.springframework.stereotype.Repository;
 import javax.transaction.Transactional;
 
 @Repository
-public interface MovieRepository extends JpaRepository<Movie,Integer> {
+public interface MovieRepository extends JpaRepository<Movie,Long> {
     Movie findByTitle(String title);
+
     @Transactional
     void deleteByTitle(String title);
 }
